requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
pytz==2018.4
